requirements.txt
fastapi
uvicorn
pillow
tensorflow
jinja2
